ref(ddm): remove calls to ddm/meta endpoint
