+++ static/app/views/ddm/widgetDetails.tsx
@@ -25,8 +25,6 @@ import useOrganization from 'sentry/utils/useOrganization';
 import {CodeLocations} from 'sentry/views/ddm/codeLocations';
 import type {FocusAreaProps} from 'sentry/views/ddm/context';
 import {useDDMContext} from 'sentry/views/ddm/context';
-import type {SamplesTableProps} from 'sentry/views/ddm/sampleTable';
-import {SampleTable} from 'sentry/views/ddm/sampleTable';
 import {extendQueryWithGroupBys} from 'sentry/views/ddm/utils';
 
 enum Tab {
@@ -78,7 +76,7 @@ interface MetricDetailsProps {
   focusArea?: FocusAreaProps;
   focusedSeries?: FocusedMetricsSeries[];
   mri?: MRI;
-  onRowHover?: SamplesTableProps['onRowHover'];
+  onRowHover?: (sampleId?: string) => void;
   op?: string;
   query?: string;
   setMetricsSamples?: React.Dispatch<
@@ -86,7 +84,6 @@ interface MetricDetailsProps {
   >;
 }
 
-// TODO: add types
 export function MetricDetails({
   mri,
   op,
@@ -152,32 +149,23 @@ export function MetricDetails({
           <TabPanels>
             <TabPanels.Item key={Tab.SAMPLES}>
               <MetricSampleTableWrapper organization={organization}>
-                {organization.features.includes('metrics-samples-list') ? (
-                  organization.features.includes('metrics-samples-list-search') ? (
-                    <SearchableMetricSamplesTable
-                      focusArea={focusArea?.selection?.range}
-                      mri={mri}
-                      onRowHover={onRowHover}
-                      op={op}
-                      query={queryWithFocusedSeries}
-                      setMetricsSamples={setMetricsSamples}
-                    />
-                  ) : (
-                    <MetricSamplesTable
-                      focusArea={focusArea?.selection?.range}
-                      mri={mri}
-                      onRowHover={onRowHover}
-                      op={op}
-                      query={queryWithFocusedSeries}
-                      setMetricsSamples={setMetricsSamples}
-                    />
-                  )
-                ) : (
-                  <SampleTable
+                {organization.features.includes('metrics-samples-list-search') ? (
+                  <SearchableMetricSamplesTable
+                    focusArea={focusArea?.selection?.range}
                     mri={mri}
-                    {...focusArea?.selection?.range}
-                    query={queryWithFocusedSeries}
                     onRowHover={onRowHover}
+                    op={op}
+                    query={queryWithFocusedSeries}
+                    setMetricsSamples={setMetricsSamples}
+                  />
+                ) : (
+                  <MetricSamplesTable
+                    focusArea={focusArea?.selection?.range}
+                    mri={mri}
+                    onRowHover={onRowHover}
+                    op={op}
+                    query={queryWithFocusedSeries}
+                    setMetricsSamples={setMetricsSamples}
                   />
                 )}
               </MetricSampleTableWrapper>
